third_party/WebKit/LayoutTests/fast/css/preferred-stylesheet-enabling-alternate.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel="stylesheet alternate" href="data:text/css,p { color:green }" title="prefer">
<link rel="stylesheet" href="data:text/css," title="prefer">
<p id="t">This text should be green</p>
<script>
    test(() => {
        assert_equals(getComputedStyle(t).color, "rgb(0, 128, 0)",
            "The rule in the alternate stylesheet did not apply to the p element.");
    }, "Test that preferred stylesheet is enabling alternate stylesheet with the same title.");
</script>
